+++ src/main/java/tornadofx/Controls.kt
@@ -96,6 +96,7 @@ val TabPane.refreshable: BooleanExpression get() {
 }
 
 inline fun <reified T : UIComponent> TabPane.contentUiComponent(): T? = selectionModel.selectedItem?.content?.uiComponent<T>()
+fun TabPane.onDelete() = contentUiComponent<UIComponent>()?.onDelete()
 fun TabPane.onSave() = contentUiComponent<UIComponent>()?.onSave()
 fun TabPane.onRefresh() = contentUiComponent<UIComponent>()?.onRefresh()
 fun TabPane.onNavigateBack() = contentUiComponent<UIComponent>()?.onNavigateBack() ?: true